LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/085-expected.txt

FAILED (This TC requires JavaScript enabled)

FAIL  scheduler: async script and slow-loading defer script assert_array_equals: expected property 0 to be "external script #2" but got "external script #1" (expected array ["external script #2", "external script #1"] got ["external script #1", "external script #2"])
